src/requirements.txt
scikit-learn
scipy
tensorboard
tensorflow-gpu==2.4.0
jupyter
pandas
Pillow
tqdm
jedi==0.17.2
matplotlib
seaborn
tensorflow_addons
numba
tensorboard_plugin_profile
scikit-image
einops
tabulate
segmentation-models
pydot
numpy

jupyterlab
jupyterlab-server
